refactor: split engine/agent_engine.py into 5 mixins
